ui: dashboard - review comments
