+++ invenio_app_ils/ui/src/invenio_app_ils/pages/backoffice/DocumentDetails/components/DocumentPendingLoans/DocumentPendingLoans.js
@@ -8,6 +8,7 @@ import {
   loanSearchQueryUrl,
   viewLoanDetailsUrl,
 } from '../../../../../common/urls';
+import { formatter } from '../../../../../common/components/ResultsTable/formatters';
 
 export default class DocumentPendingLoans extends Component {
   constructor(props) {
@@ -39,7 +40,7 @@ export default class DocumentPendingLoans extends Component {
   };
 
   prepareData() {
-    return this.props.data.map(row => loanApi.serializer.toTableView(row));
+    return this.props.data.map(row => formatter.loan.toTable(row));
   }
 
   render() {
